requirements.txt
torch
transformers
pillow
flask
numpy
requests
geoclip
pytesseract
scipy
opencv-python
overpy
